Better claim lookup
You can now look up claims by entering a date, a substring of the
provider, or a substring of the claim ID number.

New 'claim_menu' gets a list of claims, formats them, presents the
formatted strings in a menu, and then retuns the selected claim.

Several functions that dealt with claim IDs now deal directly with
claims.